Split out Billing module from ajax
